test: deduplicate fake codex fixtures
