+++ tests/common/mod.rs
@@ -160,6 +160,74 @@ pub fn tempdir_or_panic() -> TempDir {
     }
 }
 
+pub fn fake_codex_bin(temp: &Path) -> PathBuf {
+    temp.join(if cfg!(windows) { "codex.cmd" } else { "codex" })
+}
+
+#[cfg(windows)]
+pub fn install_fake_codex_launcher(_script: &Path, bin: &Path) {
+    fs::write(bin, windows_python_launcher("codex.py")).unwrap_or_else(|err| {
+        panic!(
+            "failed to install fake codex launcher {}: {err}",
+            bin.display()
+        )
+    });
+}
+
+#[cfg(not(windows))]
+pub fn install_fake_codex_launcher(script: &Path, bin: &Path) {
+    fs::copy(script, bin).unwrap_or_else(|err| {
+        panic!(
+            "failed to install fake codex launcher {} from {}: {err}",
+            bin.display(),
+            script.display()
+        )
+    });
+    make_executable(bin);
+}
+
+#[cfg(unix)]
+fn make_executable(path: &Path) {
+    use std::os::unix::fs::PermissionsExt;
+
+    let mut permissions = fs::metadata(path)
+        .unwrap_or_else(|err| panic!("failed to stat {}: {err}", path.display()))
+        .permissions();
+    permissions.set_mode(0o755);
+    fs::set_permissions(path, permissions)
+        .unwrap_or_else(|err| panic!("failed to chmod {}: {err}", path.display()));
+}
+
+#[cfg(not(unix))]
+fn make_executable(_path: &Path) {}
+
+pub fn windows_python_launcher(script_name: &str) -> String {
+    format!(
+        "@echo off\r\n\
+setlocal\r\n\
+if defined Python_ROOT_DIR if exist \"%Python_ROOT_DIR%\\python.exe\" (\r\n\
+  \"%Python_ROOT_DIR%\\python.exe\" \"%~dp0{script_name}\" %*\r\n\
+  exit /b %ERRORLEVEL%\r\n\
+)\r\n\
+if defined pythonLocation if exist \"%pythonLocation%\\python.exe\" (\r\n\
+  \"%pythonLocation%\\python.exe\" \"%~dp0{script_name}\" %*\r\n\
+  exit /b %ERRORLEVEL%\r\n\
+)\r\n\
+where python >nul 2>nul\r\n\
+if not errorlevel 1 (\r\n\
+  python \"%~dp0{script_name}\" %*\r\n\
+  exit /b %ERRORLEVEL%\r\n\
+)\r\n\
+where python3 >nul 2>nul\r\n\
+if not errorlevel 1 (\r\n\
+  python3 \"%~dp0{script_name}\" %*\r\n\
+  exit /b %ERRORLEVEL%\r\n\
+)\r\n\
+py -3 \"%~dp0{script_name}\" %*\r\n\
+exit /b %ERRORLEVEL%\r\n"
+    )
+}
+
 pub fn sample_node(id: &str, name: &str, file_path: &str) -> Node {
     Node {
         id: id.to_string(),
