+++ tests/dashboard_api_support/mod.rs
@@ -6,8 +6,9 @@ pub(crate) use std::process::Command;
 pub(crate) use std::thread;
 
 pub(crate) use crate::common::{
-    create_runtime, get_json, http_agent, pick_free_port, response_to_json, tempdir_or_panic,
-    wait_for_dashboard, EnvVarGuard, GLOBAL_DB_ENV, GLOBAL_DB_ENV_LOCK,
+    create_runtime, fake_codex_bin, get_json, http_agent, install_fake_codex_launcher,
+    pick_free_port, response_to_json, tempdir_or_panic, wait_for_dashboard, EnvVarGuard,
+    GLOBAL_DB_ENV, GLOBAL_DB_ENV_LOCK,
 };
 pub(crate) use serde_json::Value;
 pub(crate) use tempfile::TempDir;
@@ -470,69 +471,6 @@ for line in sys.stdin:
         install_fake_codex_launcher(&script_path, &bin);
         Self { _temp: temp, bin }
     }
-}
-
-fn fake_codex_bin(temp: &Path) -> PathBuf {
-    temp.join(if cfg!(windows) { "codex.cmd" } else { "codex" })
-}
-
-#[cfg(windows)]
-fn install_fake_codex_launcher(_script: &Path, bin: &Path) {
-    write_file(bin, &windows_python_launcher("codex.py"));
-}
-
-#[cfg(not(windows))]
-fn install_fake_codex_launcher(script: &Path, bin: &Path) {
-    if let Err(err) = fs::copy(script, bin) {
-        panic!(
-            "failed to install fake codex launcher {} from {}: {err}",
-            bin.display(),
-            script.display()
-        );
-    }
-    make_executable(bin);
-}
-
-#[cfg(unix)]
-pub(crate) fn make_executable(path: &Path) {
-    use std::os::unix::fs::PermissionsExt;
-
-    let mut permissions = fs::metadata(path)
-        .unwrap_or_else(|err| panic!("failed to stat {}: {err}", path.display()))
-        .permissions();
-    permissions.set_mode(0o755);
-    fs::set_permissions(path, permissions)
-        .unwrap_or_else(|err| panic!("failed to chmod {}: {err}", path.display()));
-}
-
-#[cfg(not(unix))]
-pub(crate) fn make_executable(_path: &Path) {}
-
-fn windows_python_launcher(script_name: &str) -> String {
-    format!(
-        "@echo off\r\n\
-setlocal\r\n\
-if defined Python_ROOT_DIR if exist \"%Python_ROOT_DIR%\\python.exe\" (\r\n\
-  \"%Python_ROOT_DIR%\\python.exe\" \"%~dp0{script_name}\" %*\r\n\
-  exit /b %ERRORLEVEL%\r\n\
-)\r\n\
-if defined pythonLocation if exist \"%pythonLocation%\\python.exe\" (\r\n\
-  \"%pythonLocation%\\python.exe\" \"%~dp0{script_name}\" %*\r\n\
-  exit /b %ERRORLEVEL%\r\n\
-)\r\n\
-where python >nul 2>nul\r\n\
-if not errorlevel 1 (\r\n\
-  python \"%~dp0{script_name}\" %*\r\n\
-  exit /b %ERRORLEVEL%\r\n\
-)\r\n\
-where python3 >nul 2>nul\r\n\
-if not errorlevel 1 (\r\n\
-  python3 \"%~dp0{script_name}\" %*\r\n\
-  exit /b %ERRORLEVEL%\r\n\
-)\r\n\
-py -3 \"%~dp0{script_name}\" %*\r\n\
-exit /b %ERRORLEVEL%\r\n"
-    )
 }
 
 pub(crate) async fn start_dashboard_fixture(seed_lcm: bool) -> DashboardFixture {
